[#11] Move all ui pages to corresponding packages
